(#1) Use - rather than _ in assembly name
When importing TagHelpers, you specify the name of the assembly that it
should look for them. When this template was created, I used glaa-trips,
as the name, however, this was changed to glaa_trips within the
templated files, while the assembly is still called glaa-trips.

As a result, the TagHelpers are never found. Changing the assembly name
makes everything start working again.

This was found in this Stack Overflow question:

https://stackoverflow.com/a/60489563/671491