(#6) Additional implementation
Added the ability to seriliaze metadata of album to a file, which can
then be read back. On top of this, a collection of marker locations is
being stored in a single file, to then allow the population of the
Google Maps widget.  Finally, added the necessary CSS to make the map
widget appear on the screen.