(#10) Add ability to edit metadata
Added concept of DisplayName for an album, as well as Id property. The
Id will be the generated slug for the album, which is essentially fixed,
where as the DisplayName can be changed.

When editing and deleting an album, the markers.json file is
re-generated based on current album collection.

Simple form was added to allow editing of album metadata.